Split federation tests to more evenly distribute over workers
